+++ Tests/MultipasteCoreTests/main.swift
@@ -7,5 +7,6 @@ SnippetMatcherTests.registerAll()
 SemanticVersionTests.registerAll()
 UpdateCheckerTests.registerAll()
 PasteboardAugmenterTests.registerAll()
+TabNavigationTests.registerAll()
 
 exit(Int32(runAllTests()))